Split multi-component files into one component per file
Extract CertSection from certs-view.tsx into its own file. Split
status-messages.tsx into loading-state.tsx and error-state.tsx.
Each application file now has a single component.
