+++ frontend/src/views/projects-view.tsx
@@ -3,7 +3,8 @@ import { ProjectList } from "@/components/project-list";
 import { AddProjectDialog } from "@/components/add-project-dialog";
 import { EditProjectDialog } from "@/components/edit-project-dialog";
 import { SectionHeader } from "@/components/section-header";
-import { LoadingState, ErrorState } from "@/components/status-messages";
+import { LoadingState } from "@/components/loading-state";
+import { ErrorState } from "@/components/error-state";
 import { Button } from "@/components/ui/button";
 import { useProjects } from "@/hooks/use-projects";
 import { ProjectServicesProvider } from "@/hooks/use-project-services";
